1. Users management: Modified users management layout
