+++ src/pages/Admin/Users/Register/RegisterUsers.jsx
@@ -8,6 +8,7 @@ import {
   AdminPanelSettingsRounded,
   BadgeRounded,
   Email,
+  PersonAddRounded,
   School,
 } from "@mui/icons-material";
 
@@ -52,12 +53,7 @@ import AppButton from "../../../../components/Button/AppButton";
 import Loader from "../../../../components/Loader/Loader";
 import Slack from "../../Cohorts/TableRender/Slack";
 import UserAvatarRender from "../RegisterOnCohort/TableRenders/UserAvatarRender";
-/*
-    ==========================
-    =     AUX VARIABLES      =
-    ==========================
-*/
-const rolesList = ["Admin", "Mentor", "Student"];
+import AddUser from "./Actions/AddUser";
 
 const RegisterUsers = () => {
   /*
@@ -69,44 +65,13 @@ const RegisterUsers = () => {
   const navigate = useNavigate();
   const location = useLocation();
   const { setAuth } = useAuth();
-
-  /*
-    ==========================
-    =         STATES         =
-    ==========================
-  */
   // Fetched data states:
   const [users, setUsers] = useState([]);
-  console.log(users);
   const [cohorts, setCohorts] = useState([]);
-  // From states
-  const [userRoles, setUserRoles] = useState([]);
-  const [cohortsValueSelected, setCohortsValueSelected] = useState({
-    id: "",
-    cohort: "",
-  });
-  const [cohortsInputValueSelected, setCohortsInputValueSelected] =
-    useState("");
+
+  console.log(users);
   const [loading, setLoading] = useState(false);
-  const [formError, setFormError] = useState({
-    userNameError: {
-      error: false,
-      errorMessage: "Please enter a valid name",
-    },
-    userEmailError: {
-      error: false,
-      errorMessage: "Please enter a valid e-mail address",
-    },
-    userCohortError: {
-      error: false,
-      errorMessage: "Please add a cohort for this user",
-    },
-    userRolesError: {
-      error: true,
-      errorMessage: "Please select a role for this user",
-    },
-  });
-  const [reset, setReset] = useState(false);
+  const [openAddUserDialog, setOpenAddUserDialog] = useState(false);
   /*
     ==========================
     =      AUX VARIABLES     =
@@ -114,7 +79,13 @@ const RegisterUsers = () => {
   */
   const columns = [
     { field: "id", headerName: "ID", minWidth: 100, maxWidth: 130, flex: 1 },
-    { field: "slackId", headerName: "ID", minWidth: 100, maxWidth: 130, flex: 1 },
+    {
+      field: "slackId",
+      headerName: "ID",
+      minWidth: 100,
+      maxWidth: 130,
+      flex: 1,
+    },
     {
       field: "slack",
       headerName: "Created on:",
@@ -139,21 +110,24 @@ const RegisterUsers = () => {
       flex: 1,
       valueGetter: (params) => params,
       renderCell: (params) => (
-        <UserAvatarRender name={params.row.userName} avatarUrl={params.row.userAvatar}></UserAvatarRender>
+        <UserAvatarRender
+          name={params.row.userName}
+          avatarUrl={params.row.userAvatar}
+        ></UserAvatarRender>
       ),
     },
     {
       field: "userName",
       headerName: "Name:",
-      minWidth: 225,
-      maxWidth: 225,
+      minWidth: 200,
+      maxWidth: 200,
       flex: 1,
     },
     {
       field: "userEmail",
       headerName: "E-mail:",
-      minWidth: 300,
-      maxWidth: 300,
+      minWidth: 225,
+      maxWidth: 225,
       flex: 1,
     },
     {
@@ -161,8 +135,8 @@ const RegisterUsers = () => {
       headerName: "Cohort: ",
       sortable: false,
       disableColumnMenu: true,
-      minWidth: 350,
-      maxWidth: 350,
+      minWidth: 175,
+      maxWidth: 175,
       flex: 1,
       valueGetter: (params) => params,
       renderCell: (params) => (
@@ -185,8 +159,8 @@ const RegisterUsers = () => {
       headerName: "Status: ",
       sortable: false,
       disableColumnMenu: true,
-      minWidth: 100,
-      maxWidth: 100,
+      minWidth: 95,
+      maxWidth: 95,
       flex: 1,
       valueGetter: (params) => params,
       renderCell: (params) => (
@@ -274,7 +248,6 @@ const RegisterUsers = () => {
         };
       });
       setCohorts(formattedCohorts);
-      setCohortsValueSelected(formattedCohorts[0]);
     } catch (error) {
       if (error.response.status === 403) {
         console.error(error);
@@ -297,93 +270,6 @@ const RegisterUsers = () => {
     setLoading(false);
   };
 
-  /*
-    ==========================
-    =   HANDLER FUNCTIONS    =
-    ==========================
-  */
-  // User name:
-  const handleUserNameError = (inputError) => {
-    setFormError((prevState) => ({
-      ...prevState,
-      userNameError: {
-        ...prevState.userNameError,
-        error: inputError,
-      },
-    }));
-  };
-
-  //User email
-  const handleUserEmailError = (inputError) => {
-    setFormError((prevState) => ({
-      ...prevState,
-      userEmailError: {
-        ...prevState.userEmailError,
-        error: inputError,
-      },
-    }));
-  };
-
-  //User cohorts
-  const handleValueSelectedChange = (newValue) => {
-    setCohortsValueSelected(newValue);
-  };
-
-  // onSelect Role:
-  const handleOnSelectRole = (selectedRoleName) => {
-    setUserRoles(selectedRoleName);
-  };
-
-  // Form submit:
-  const handleRegisterOnCohortSubmit = async (event) => {
-    event.preventDefault();
-    const formattedUserRegistration = {
-      users: [
-        {
-          name: event.target.userName.value.trim(),
-          email: event.target.userEmail.value.trim(),
-          role: userRoles.map((role) => role.toLowerCase()),
-        },
-      ],
-      cohort: cohortsValueSelected.id,
-    };
-    const errors = Object.values(formError);
-    try {
-      setLoading(true);
-      if (!errors.some((error) => error.error === true)) {
-        const response = await axiosPrivate.post(
-          "auth/register",
-          formattedUserRegistration
-        );
-        console.log(response);
-        if (response.data.users.length > 0) {
-          setUsers((prevState) => [
-            ...prevState,
-            {
-              id: response.data.users[0]._id,
-              userName: response.data.users[0].name,
-              userEmail: response.data.users[0].email,
-              userCohort: [cohortsValueSelected],
-              userRole: response.data.users[0].role,
-              userActivatedStatus: response.data.users[0].isActivated,
-            },
-          ]);
-          setReset(true);
-          setCohortsValueSelected(cohorts[0]);
-          setUserRoles([]);
-        } else if (response.data.errors.length > 0) {
-          console.error(response.data.errors);
-        }
-      } else {
-        console.error(
-          "There is an error preventing the form submission: check that your entires are correctly validated"
-        );
-      }
-    } catch (error) {
-      console.error(error);
-    }
-    setLoading(false);
-  };
   /* 
       ==========================
       =        EFFECTS         =
@@ -394,28 +280,8 @@ const RegisterUsers = () => {
     fetchCohorts();
   }, []);
 
-  useEffect(() => {
-    setFormError((prevState) => ({
-      ...prevState,
-      userCohortError: {
-        ...prevState.userCohortError,
-        error: cohortsValueSelected === null ? true : false,
-      },
-    }));
-  }, [cohortsValueSelected]);
-
-  useEffect(() => {
-    setFormError((prevState) => ({
-      ...prevState,
-      userRolesError: {
-        ...prevState.userRolesError,
-        error: userRoles.length === 0 ? true : false,
-      },
-    }));
-  }, [userRoles]);
-
   return (
-    <Container maxWidth="xl">
+    <Container maxWidth="lg">
       <Paper
         elevation={3}
         sx={{
@@ -447,7 +313,6 @@ const RegisterUsers = () => {
           Users Management{" "}
         </Typography>
         <Box
-          component={"form"}
           sx={{
             display: "flex",
             flexDirection: "column",
@@ -455,122 +320,18 @@ const RegisterUsers = () => {
             alignItems: "center",
             width: "100%",
           }}
-          autoComplete="off"
-          onSubmit={handleRegisterOnCohortSubmit}
         >
           <div className={styles.formContainer}>
-            <AuthFormControl width="75%">
-              <Box
-                sx={{
-                  display: "flex",
-                  flexDirection: "column",
-                  justifyContent: "center",
-                }}
-              >
-                <BadgeRounded fontSize="large"></BadgeRounded>
-                <br></br>
-                <br></br>
-              </Box>
-              <FormTextField
-                required
-                type="text"
-                label="Name:"
-                name="userName"
-                isFocused={true}
+            <AuthFormControl width="30%">
+              <AppButton
+                text={"Add new user"}
+                type="button"
                 width="100%"
-                variant="light"
-                regex={/^(?!\s)(.{3,})(?<!\s)$/}
-                onHandleError={handleUserNameError}
-                errorMessage={"Please enter a valid name"}
-                reset={reset}
-              ></FormTextField>
-              <Box
-                sx={{
-                  display: "flex",
-                  flexDirection: "column",
-                  justifyContent: "center",
-                }}
+                handlerFunction={() => setOpenAddUserDialog(true)}
               >
-                <Email fontSize="large"></Email>
-                <br></br>
-                <br></br>
-              </Box>
-              <FormTextField
-                required
-                type="text"
-                label="E-mail:"
-                name="userEmail"
-                isFocused={false}
-                width="100%"
-                variant="light"
-                regex={/^[^\s@]+@[^\s@]+\.[^\s@]+$/}
-                onHandleError={handleUserEmailError}
-                errorMessage={"Please enter a valid e-mail address"}
-                reset={reset}
-              ></FormTextField>
+                <PersonAddRounded fontSize="large"></PersonAddRounded>
+              </AppButton>
             </AuthFormControl>
-            <AuthFormControl width="75%">
-              <Box
-                sx={{
-                  display: "flex",
-                  flexDirection: "column",
-                  justifyContent: "center",
-                }}
-              >
-                <School fontSize="large" />
-                <br></br>
-                <br></br>
-              </Box>
-              <AuthFormControl width="100%" isNested={true}>
-                <FormAutocomplete
-                  multiple={false}
-                  value={cohortsValueSelected}
-                  computedIdProperty={"id"}
-                  computedProperty={"cohort"}
-                  onHandleSelectedValueChange={handleValueSelectedChange}
-                  inputValue={cohortsInputValueSelected}
-                  onHandleInputValueChange={setCohortsInputValueSelected}
-                  options={cohorts}
-                  error={formError.userCohortError}
-                  variant={"light"}
-                ></FormAutocomplete>
-              </AuthFormControl>
-            </AuthFormControl>
-            <AuthFormControl width="75%">
-              <Box
-                sx={{
-                  display: "flex",
-                  flexDirection: "column",
-                  justifyContent: "center",
-                }}
-              >
-                <AdminPanelSettingsRounded fontSize="large" />
-                <br></br>
-              </Box>
-              <AuthFormControl
-                width="100%"
-                isNested={true}
-                error={formError.userRolesError.error}
-              >
-                <FormSelect
-                  id={"userRoles"}
-                  name={"userRoles"}
-                  label={"Roles:"}
-                  selectValue={userRoles}
-                  onSelectValue={handleOnSelectRole}
-                  list={rolesList}
-                  variant={"light"}
-                  multiple={true}
-                  error={formError.userRolesError}
-                ></FormSelect>
-              </AuthFormControl>
-            </AuthFormControl>
-            <AppButton
-              text={"Add new user"}
-              type="submit"
-              width="25%"
-              handlerFunction={() => {}}
-            />
           </div>
         </Box>
         {loading ? (
@@ -585,6 +346,15 @@ const RegisterUsers = () => {
             sortType={"asc"}
           />
         )}
+        {openAddUserDialog ? (
+          <AddUser
+            open={openAddUserDialog}
+            cohorts={cohorts}
+            handleOpen={setOpenAddUserDialog}
+            onHandleUsers={setUsers}
+            onLoading={setLoading}
+          ></AddUser>
+        ) : null}
       </Paper>
     </Container>
   );
